backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic-settings
python-jose[cryptography]
